cleanup(storage): Factor out test constants into common location
Fixes: #3022
