+++ google/cloud/storage/testing/write_base64.cc
@@ -1,0 +1,36 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "google/cloud/storage/testing/write_base64.h"
+#include "google/cloud/storage/internal/openssl_util.h"
+#include <fstream>
+
+namespace google {
+namespace cloud {
+namespace storage {
+namespace testing {
+void WriteBase64AsBinary(std::string const& filename, char const* data) {
+  std::ofstream os(filename, std::ios::binary);
+  os.exceptions(std::ios::badbit);
+  auto bytes = internal::Base64Decode(data);
+  for (unsigned char c : bytes) {
+    os << c;
+  }
+  os.close();
+}
+
+}  // namespace testing
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
